+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - New: Add new `is_locked` property to show if a page is locked or not.
 - Fix: Pagination when appending block children no longer leads to repeated appendings.
+- Fix: Assembling a hierarchy of blocks offline and appending it will result in automatic batching
+       adhering to the Notion API limits, issue #119, #120.
 
 ## Version 0.9.1, 2025-09-03